+++ vm/main.tf
@@ -1,6 +1,5 @@
 # ./vm/main.tf - Azure VM resources for WireGuard
 
-# Public IP Address for the VM
 resource "azurerm_public_ip" "wireguard_public_ip" {
   name                = "${var.vm_name}-ip"
   resource_group_name = var.resource_group_name
@@ -11,7 +10,6 @@ resource "azurerm_public_ip" "wireguard_public_ip" {
   tags = var.tags
 }
 
-# Network Interface for the VM
 resource "azurerm_network_interface" "wireguard_nic" {
   name                = "${var.vm_name}-nic"
   resource_group_name = var.resource_group_name
@@ -27,7 +25,6 @@ resource "azurerm_network_interface" "wireguard_nic" {
   tags = var.tags
 }
 
-# Ubuntu VM for WireGuard
 resource "azurerm_linux_virtual_machine" "wireguard_vm" {
   name                = var.vm_name
   resource_group_name = var.resource_group_name
@@ -38,7 +35,6 @@ resource "azurerm_linux_virtual_machine" "wireguard_vm" {
     azurerm_network_interface.wireguard_nic.id,
   ]
 
-  # Use SSH key for authentication (more secure than password)
   admin_ssh_key {
     username   = var.admin_username
     public_key = var.ssh_public_key
@@ -50,7 +46,6 @@ resource "azurerm_linux_virtual_machine" "wireguard_vm" {
     disk_size_gb         = var.os_disk_size_gb
   }
 
-  # Ubuntu Server image
   source_image_reference {
     publisher = var.ubuntu_version.publisher
     offer     = var.ubuntu_version.offer
@@ -58,12 +53,10 @@ resource "azurerm_linux_virtual_machine" "wireguard_vm" {
     version   = var.ubuntu_version.version
   }
 
-  # Enable system-assigned managed identity
   identity {
     type = "SystemAssigned"
   }
   
-  # Add custom data for cloud-init (optional, for basic setup)
   custom_data = base64encode(<<-EOF
     #!/bin/bash
     # Update and install basic utilities